JSON & D3.Simulation/Force
Added new code to crunch the package data from the CDA generated XML,
create a nodes and links colelction that we can use for D3 force and
generate the  base model. I still need to tweak it since it does not
have a dynamic canvas to see all of the data and it still needs more
styling.